findomain/requirements.txt
Flask==2.2.2
Flask-APScheduler==1.12.4
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
marshmallow-sqlalchemy==0.28.1
python-dotenv==0.21.0
requests==2.28.1